+++ src/hint_processor/builtin_hint_processor/ec_recover.rs
@@ -1,4 +1,7 @@
-use super::secp::{bigint_utils::BigInt3, secp_utils::pack};
+use num_integer::Integer;
+
+use super::secp::bigint_utils::BigInt3;
+use super::secp::secp_utils::bigint3_pack;
 use crate::stdlib::{collections::HashMap, prelude::*};
 use crate::{
     hint_processor::hint_processor_definition::HintReference,
@@ -7,7 +10,7 @@ use crate::{
     types::exec_scope::ExecutionScopes,
     vm::{errors::hint_errors::HintError, vm_core::VirtualMachine},
 };
-use num_integer::Integer;
+use num_bigint::BigInt;
 
 /* Implements Hint:
 %{
@@ -26,9 +29,9 @@ pub fn ec_recover_divmod_n_packed(
     ids_data: &HashMap<String, HintReference>,
     ap_tracking: &ApTracking,
 ) -> Result<(), HintError> {
-    let n = pack(BigInt3::from_var_name("n", vm, ids_data, ap_tracking)?);
-    let x = pack(BigInt3::from_var_name("x", vm, ids_data, ap_tracking)?).mod_floor(&n);
-    let s = pack(BigInt3::from_var_name("s", vm, ids_data, ap_tracking)?).mod_floor(&n);
+    let n = bigint3_pack(BigInt3::from_var_name("n", vm, ids_data, ap_tracking)?);
+    let x = bigint3_pack(BigInt3::from_var_name("x", vm, ids_data, ap_tracking)?).mod_floor(&n);
+    let s = bigint3_pack(BigInt3::from_var_name("s", vm, ids_data, ap_tracking)?).mod_floor(&n);
 
     let value = div_mod(&x, &s, &n);
     exec_scopes.insert_value("value", value.clone());
@@ -52,12 +55,58 @@ pub fn ec_recover_sub_a_b(
     ids_data: &HashMap<String, HintReference>,
     ap_tracking: &ApTracking,
 ) -> Result<(), HintError> {
-    let a = pack(BigInt3::from_var_name("a", vm, ids_data, ap_tracking)?);
-    let b = pack(BigInt3::from_var_name("b", vm, ids_data, ap_tracking)?);
+    let a = bigint3_pack(BigInt3::from_var_name("a", vm, ids_data, ap_tracking)?);
+    let b = bigint3_pack(BigInt3::from_var_name("b", vm, ids_data, ap_tracking)?);
 
     let value = a - b;
     exec_scopes.insert_value("value", value.clone());
     exec_scopes.insert_value("res", value);
+    Ok(())
+}
+
+/* Implements Hint:
+%{
+    from starkware.cairo.common.cairo_secp.secp_utils import pack
+    from starkware.python.math_utils import div_mod, safe_div
+
+    a = pack(ids.a, PRIME)
+    b = pack(ids.b, PRIME)
+    product = a * b
+    m = pack(ids.m, PRIME)
+
+    value = res = product % m
+%}
+ */
+pub fn ec_recover_product_mod(
+    vm: &mut VirtualMachine,
+    exec_scopes: &mut ExecutionScopes,
+    ids_data: &HashMap<String, HintReference>,
+    ap_tracking: &ApTracking,
+) -> Result<(), HintError> {
+    let a = bigint3_pack(BigInt3::from_var_name("a", vm, ids_data, ap_tracking)?);
+    let b = bigint3_pack(BigInt3::from_var_name("b", vm, ids_data, ap_tracking)?);
+    let m = bigint3_pack(BigInt3::from_var_name("m", vm, ids_data, ap_tracking)?);
+
+    let product = a * b;
+    let value = product.mod_floor(&m);
+    exec_scopes.insert_value("product", product);
+    exec_scopes.insert_value("m", m);
+    exec_scopes.insert_value("value", value.clone());
+    exec_scopes.insert_value("res", value);
+    Ok(())
+}
+
+/* Implements Hint:
+%{
+    value = k = product // m
+%}
+ */
+pub fn ec_recover_product_div_m(exec_scopes: &mut ExecutionScopes) -> Result<(), HintError> {
+    let product: &BigInt = exec_scopes.get_ref("product")?;
+    let m: &BigInt = exec_scopes.get_ref("m")?;
+    let value = product.div_floor(m);
+    exec_scopes.insert_value("k", value.clone());
+    exec_scopes.insert_value("value", value);
     Ok(())
 }
 
@@ -69,10 +118,7 @@ mod tests {
     use crate::hint_processor::builtin_hint_processor::hint_code;
     use crate::hint_processor::hint_processor_definition::HintReference;
     use crate::utils::test_utils::*;
-    use crate::vm::errors::memory_errors::MemoryError;
     use crate::vm::vm_core::VirtualMachine;
-    use crate::vm::vm_memory::memory::Memory;
-    use crate::vm::vm_memory::memory_segments::MemorySegmentManager;
     use crate::{
         any_box,
         hint_processor::{
@@ -81,7 +127,7 @@ mod tests {
             },
             hint_processor_definition::HintProcessor,
         },
-        types::{exec_scope::ExecutionScopes, relocatable::MaybeRelocatable},
+        types::exec_scope::ExecutionScopes,
     };
 
     #[cfg(target_arch = "wasm32")]
@@ -156,6 +202,73 @@ mod tests {
         check_scope!(
             &exec_scopes,
             [("value", BigInt::from(75)), ("res", BigInt::from(75))]
+        );
+    }
+
+    #[test]
+    #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+    fn run_ec_recover_product_mod_ok() {
+        let mut vm = vm!();
+        let mut exec_scopes = ExecutionScopes::new();
+
+        vm.run_context.fp = 8;
+        let ids_data = non_continuous_ids_data![("a", -8), ("b", -5), ("m", -2)];
+
+        vm.segments = segments![
+            //a
+            ((1, 0), 60),
+            ((1, 1), 0),
+            ((1, 2), 0),
+            //b
+            ((1, 3), 2),
+            ((1, 4), 0),
+            ((1, 5), 0),
+            //m
+            ((1, 6), 100),
+            ((1, 7), 0),
+            ((1, 8), 0)
+        ];
+
+        assert!(run_hint!(
+            vm,
+            ids_data,
+            hint_code::EC_RECOVER_PRODUCT_MOD,
+            &mut exec_scopes
+        )
+        .is_ok());
+
+        check_scope!(
+            &exec_scopes,
+            [
+                ("value", BigInt::from(20)),
+                ("res", BigInt::from(20)),
+                ("product", BigInt::from(120)),
+                ("m", BigInt::from(100))
+            ]
+        );
+    }
+
+    #[test]
+    #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+    fn run_ec_recover_product_div_m_ok() {
+        let mut vm = vm!();
+        let mut exec_scopes = ExecutionScopes::new();
+        exec_scopes.insert_value("product", BigInt::from(250));
+        exec_scopes.insert_value("m", BigInt::from(100));
+
+        let ids_data = ids_data!["none"];
+
+        assert!(run_hint!(
+            vm,
+            ids_data,
+            hint_code::EC_RECOVER_PRODUCT_DIV_M,
+            &mut exec_scopes
+        )
+        .is_ok());
+
+        check_scope!(
+            &exec_scopes,
+            [("value", BigInt::from(2)), ("k", BigInt::from(2))]
         );
     }
 }