+++ src/hint_processor/builtin_hint_processor/secp/secp_utils.rs
@@ -8,7 +8,7 @@ use lazy_static::lazy_static;
 use num_bigint::{BigInt, BigUint};
 use num_traits::Zero;
 
-use super::bigint_utils::BigInt3;
+use super::bigint_utils::Uint384;
 
 // Constants in package "starkware.cairo.common.cairo_secp.constants".
 pub const BASE_86: &str = "starkware.cairo.common.cairo_secp.constants.BASE";
@@ -25,6 +25,12 @@ lazy_static! {
     //SECP_P = 2**256 - 2**32 - 2**9 - 2**8 - 2**7 - 2**6 - 2**4 - 1
     pub(crate) static ref SECP_P: BigInt = BigInt::from_str(
         "115792089237316195423570985008687907853269984665640564039457584007908834671663"
+    )
+    .unwrap();
+
+    //SECP_P_V2 = 2**255-19
+    pub(crate) static ref SECP_P_V2: BigInt = BigInt::from_str(
+        "57896044618658097711785492504343953926634992332820282019728792003956564819949"
     )
     .unwrap();
     // BASE = 2**86
@@ -62,7 +68,7 @@ Takes a 256-bit integer and returns its canonical representation as:
 d0 + BASE * d1 + BASE**2 * d2,
 where BASE = 2**86.
 */
-pub fn split(integer: &num_bigint::BigUint) -> Result<[num_bigint::BigUint; 3], HintError> {
+pub fn bigint3_split(integer: &num_bigint::BigUint) -> Result<[num_bigint::BigUint; 3], HintError> {
     let mut canonical_repr: [num_bigint::BigUint; 3] = Default::default();
     let mut num = integer.clone();
     for item in &mut canonical_repr {
@@ -81,8 +87,8 @@ Takes an UnreducedFelt2523 struct which represents a triple of limbs (d0, d1, d2
 elements and reconstructs the corresponding 256-bit integer (see split()).
 Note that the limbs do not have to be in the range [0, BASE).
 */
-pub(crate) fn pack(num: BigInt3) -> num_bigint::BigInt {
-    let limbs = vec![num.d0, num.d1, num.d2];
+pub(crate) fn bigint3_pack(num: Uint384) -> num_bigint::BigInt {
+    let limbs = [num.d0, num.d1, num.d2];
     #[allow(deprecated)]
     limbs
         .into_iter()
@@ -110,22 +116,22 @@ mod tests {
         let mut constants = HashMap::new();
         constants.insert(BASE_86.to_string(), Felt252::one() << 86_usize);
 
-        let array_1 = split(&BigUint::zero());
+        let array_1 = bigint3_split(&BigUint::zero());
         #[allow(deprecated)]
-        let array_2 = split(
+        let array_2 = bigint3_split(
             &bigint!(999992)
                 .to_biguint()
                 .expect("Couldn't convert to BigUint"),
         );
         #[allow(deprecated)]
-        let array_3 = split(
+        let array_3 = bigint3_split(
             &bigint_str!("7737125245533626718119526477371252455336267181195264773712524553362")
                 .to_biguint()
                 .expect("Couldn't convert to BigUint"),
         );
         //TODO, Check SecpSplitutOfRange limit
         #[allow(deprecated)]
-        let array_4 = split(
+        let array_4 = bigint3_split(
             &bigint_str!(
                 "773712524553362671811952647737125245533626718119526477371252455336267181195264"
             )
@@ -175,7 +181,7 @@ mod tests {
     #[test]
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
     fn secp_pack() {
-        let pack_1 = pack(BigInt3 {
+        let pack_1 = bigint3_pack(Uint384 {
             d0: Cow::Borrowed(&Felt252::new(10_i32)),
             d1: Cow::Borrowed(&Felt252::new(10_i32)),
             d2: Cow::Borrowed(&Felt252::new(10_i32)),
@@ -185,7 +191,7 @@ mod tests {
             bigint_str!("59863107065073783529622931521771477038469668772249610")
         );
 
-        let pack_2 = pack(BigInt3 {
+        let pack_2 = bigint3_pack(Uint384 {
             d0: Cow::Borrowed(&felt_str!("773712524553362")),
             d1: Cow::Borrowed(&felt_str!("57408430697461422066401280")),
             d2: Cow::Borrowed(&felt_str!("1292469707114105")),